Adding a meson option to not build with cLaTeXMath
By adding -Dclatexmath=false, this patch allows one to not build with
cLaTeXMath using meson. The option defaults to true to avoid breaking
current builds of NoteKit. Another good approach would be to add a latex
option with would accept ["none", "clatexmath", "lasem"] as options.
lasem, though, no longer seems to be a priority nor a commonly choosed
option to compile with by distributions maintainers. And since the cmake
build backend will continue to exist due to its usage by upstream, and
is the recommended build tool, adding this option does not currently
seems to be of high importance, as the meson build backend is mainly
used by distribution maintainers.