added foreground colored stroke/pen support
When the alpha of a stroke is -INFINITY, it'll render the stroke with
the foreground color from the GtkStyleContext.

The black (0.0, 0.0, 0.0, 1.0) default pen color has been replaced with
a new button that is able to select that particular "color".
If a user has modified his toolbar colors, he'll have to drop/change
that color manually.